pyproject.toml
[tool.poetry]

name = "ballotbuddies"
version = "0.0"
description = "Track your friends' voter registration status."
authors = ["Citizen Labs <info@citizenlabs.org>"]

[tool.poetry.dependencies]

python = "^3.10"

# Django
django = "^3.2.8"
django-cors-headers = "*"
django-grappelli = "*"
django-multi-email-field = "*"
crispy-bootstrap5 = "*"

# Storage
dj-database-url = "~0.4.2"
psycopg2-binary = "^2.6"
django-redis = "^4.12.1"

# Utilities
minilog = "^2.0"
requests = "^2.26"

# Deployment
whitenoise = "^5.0.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

# Formatting
black = "=21.8b0"
isort = "=5.9.3"

# Linters
pylint = "^2.4.3"
mypy = "~0.790"

# Testing
pytest = "^6.2.5"
pytest-django = "*"
pytest-describe = "^1.0"
pytest-expecter = "^2.0"
pytest-randomly = "*"
pytest-cov = "^2.8.0"
freezegun = "*"
coverage = "^5.0"
coveragespace = "^4.0"

# Server
django-livereload-server = "~0.3.2"
honcho = { git = "https://github.com/nickstenning/honcho", rev = "b4de1aa0504d41c0f05efbb04f63c624fccded81" }

# E2E
splinter = "~0.15.0"
selenium = "<4" # https://github.com/cobrateam/splinter/issues/929

# Tooling
ipython = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
sniffer = "*"

[tool.isort]

profile = "black"

known_django = "django"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
